[refactor/components]: transfer components to layouts folder
