+++ src/app/app-routing.module.ts
@@ -4,11 +4,11 @@ import { RouterModule, Routes } from '@angular/router';
 const routes: Routes = [
   {
     path: '',
-    loadChildren: () => import('./home/home.module').then(m => m.HomeModule),
+    loadChildren: () => import('./layouts/home/home.module').then((m) => m.HomeModule),
   },
   {
     path: 'quiz',
-    loadChildren: () => import('./quiz/quiz.module').then(m => m.QuizModule),
+    loadChildren: () => import('./layouts/quiz/quiz.module').then((m) => m.QuizModule),
   },
 ];
