+++ src/app/layouts/quiz/components/question-card/question-card.component.spec.ts
