+++ src/app/layouts/quiz/quiz-result/quiz-result.component.ts
