+++ src/app/layouts/quiz/quiz-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { QuizComponent } from './quiz.component';
 import { QuizResultComponent } from './quiz-result/quiz-result.component';
-import { QuizResultGuardGuard } from '../guards/quiz-result-guard.guard';
+import { QuizResultGuardGuard } from '../../guards/quiz-result-guard.guard';
 
 const routes: Routes = [
   {
