+++ src/app/layouts/quiz/quiz.module.ts
@@ -4,14 +4,14 @@ import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { MatIconModule } from '@angular/material/icon';
 import { QuizComponent } from './quiz.component';
 import { QuizRoutingModule } from './quiz-routing.module';
-import { SharedModule } from '../shared/shared.module';
+import { SharedModule } from '../../shared/shared.module';
 import { QuestionCardComponent } from './components/question-card/question-card.component';
 import { QuizResultComponent } from './quiz-result/quiz-result.component';
 import { StoreModule } from '@ngrx/store';
-import { quizFeature } from '../state/quiz/quiz.feature';
+import { quizFeature } from '../../state/quiz/quiz.feature';
 import { EffectsModule } from '@ngrx/effects';
-import { QuizEffects } from '../state/quiz/quiz.effects';
-import { NumberToAlphabetPipe } from '../pipes/number-to-alphabet.pipe';
+import { QuizEffects } from '../../state/quiz/quiz.effects';
+import { NumberToAlphabetPipe } from '../../pipes/number-to-alphabet.pipe';
 
 @NgModule({
   declarations: [
